+++ sdk/networking/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial internal package: Dio client, internet reachability, typed remote failures.